// src/services/service.interface.ts
export interface ISPLists<T> {
  value: T[];
}

export interface ISContacList {
  Title: string;
  CompanyName: string;
}

export interface IContactList {
  Attachments: boolean;
  AuthorId: number;
  CellPhone: string;
  Comments: string;
  Company: string;
  ComplianceAssetId: number;
  ContentTypeId: string;
  Created: string;
  EditorId: number;
  Email: string;
  FileSystemObjectType: number;
  FirstName: string;
  FullName: string;
  GUID: string;
  HomePhone: string;
  ID: number;
  Id: number;
  JobTitle: string;
  Modified: string;
  OData__UIVersion: string;
  ServerRedirectedEmbedUri: string;
  ServerRedirectedEmbedUrl: string;
  Title: string;
  WebPage: string;
  WorkAddress: string;
  WorkCity: string;
  WorkCountry: string;
  WorkFax: string;
  WorkPhone: string;
  WorkState: string;
  WorkZip: string;
  Salary: number;
}